Remove cast Size to uint64_t

+++ source/device/switch.cpp
@@ -8,8 +8,7 @@
 namespace sim {
 
 Switch::Switch(Id a_id, ECN&& a_ecn)
-    : RoutingModule(a_id),
-      m_ecn(std::move(a_ecn)) {}
+    : RoutingModule(a_id), m_ecn(std::move(a_ecn)) {}
 
 bool Switch::notify_about_arrival(Time arrival_time) {
     return m_process_scheduler.notify_about_arriving(arrival_time,
@@ -29,7 +28,7 @@ Time Switch::process() {
 
     // requests queue size here to consider processing packet
     float ingress_queue_filling = link->get_to_ingress_queue_size() /
-                               (float)link->get_max_to_ingress_queue_size();
+                                  link->get_max_to_ingress_queue_size();
     std::optional<Packet> optional_packet = link->get_packet();
     if (!optional_packet.has_value()) {
         LOG_WARN("No packet in link");
@@ -56,7 +55,7 @@ Time Switch::process() {
     if (packet.ecn_capable_transport) {
         float egress_queue_filling =
             next_link->get_from_egress_queue_size() /
-            (float)next_link->get_max_from_egress_buffer_size();
+            next_link->get_max_from_egress_buffer_size();
         if (m_ecn.get_congestion_mark(ingress_queue_filling) ||
             m_ecn.get_congestion_mark(egress_queue_filling)) {
             packet.congestion_experienced = true;
